FFWEB-2168: refactoring, remove useless classes
